third_party/WebKit/LayoutTests/fast/mediastream/RTCPeerConnection-createOffer.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests RTCPeerConnection createOffer.");

// Note: createOffer() calls in the test runner are successful if the
// voiceActivityDetection and iceRestart options are passed with a value of true.
// In all other cases, createOffer() fails in the test runner.

var pc = null;

function dummy() {}

function requestFailed2(reason)
{
    testPassed('requestFailed was called.');

    errorReason = reason;
    shouldBe('errorReason', '"TEST_ERROR"');

    shouldNotThrow("pc.createOffer(dummy, dummy);");
    shouldNotThrow("pc.createOffer(dummy, dummy, {});");
    shouldNotThrow("pc.createOffer(dummy, dummy, {voiceActivityDetection:false});");
    shouldNotThrow("pc.createOffer(dummy, dummy, {iceRestart:true});");
    shouldNotThrow("pc.createOffer(dummy, dummy, {voiceActivityDetection:false, iceRestart:true});");

    // Deprecated.
    shouldThrow("pc.createOffer(dummy, dummy, {offerToReceiveVideo:-1, offerToReceiveAudio:0});");
    shouldThrow("pc.createOffer(dummy, dummy, {offerToReceiveVideo:0, offerToReceiveAudio:-1});");
    shouldNotThrow("pc.createOffer(dummy, dummy, {offerToReceiveVideo:1, offerToReceiveAudio:0, voiceActivityDetection:false, iceRestart:true});");

    finishJSTest();
}

function requestSucceeded2(sd)
{
    testFailed('requestSucceeded was called.');
    finishJSTest();
}

function requestFailed1()
{
    testFailed('requestFailed was called.');
    finishJSTest();
}

function requestSucceeded1(sd)
{
    testPassed('requestSucceeded was called.');

    sessionDescription = sd;
    shouldBe('sessionDescription.type', '"offer"');

    pc.createOffer(requestSucceeded2, requestFailed2);
}

pc = new webkitRTCPeerConnection(null);
pc.createOffer(requestSucceeded1, requestFailed1, {voiceActivityDetection:true, iceRestart:true});

window.jsTestIsAsync = true;
window.successfullyParsed = true;
</script>
</body>
</html>
